Adding type cast in Str_compare - strp.c

+++ strp.c
@@ -70,7 +70,7 @@ int Str_compare(const char *s1, const char *s2) {
         s2++;
     }
 
-    return *s1 - *s2;
+    return (int) *s1 - *s2;
 }
 
 /* A pointer to the first occurrence in the string pointed to by s1 of
